+++ ogsr_engine/Luabind/src/function.cpp
@@ -22,6 +22,7 @@
 #include "stdafx.h"
 
 #include <luabind/luabind.hpp>
+#include <luabind/detail/find_best_match.hpp>
 
 namespace luabind { namespace detail { namespace free_functions {
 
@@ -64,7 +65,7 @@ namespace luabind { namespace detail { namespace free_functions {
             ret = find_best_match(
                 L
               , &rep->overloads().front()
-              , (int)rep->overloads().size()
+              , rep->overloads().size()
               , sizeof(overload_rep)
               , ambiguous
               , min_match
@@ -112,7 +113,7 @@ namespace luabind { namespace detail { namespace free_functions {
                 find_exact_match(
                     L
                   , &rep->overloads().front()
-                  , (int)rep->overloads().size()
+                  , rep->overloads().size()
                   , sizeof(overload_rep)
                   , min_match
                   , num_params
